Extract InkedContainer as custom widget
